Refactored so that syntax highlight replaces error highlighting.
