+++ python/nammu/controller/AtfAreaController.py
@@ -18,7 +18,11 @@ along with Nammu.  If not, see <http://www.gnu.org/licenses/>.
 '''
 
 from javax.swing.undo import CannotUndoException, CannotRedoException
+
 from ..view.AtfAreaView import AtfAreaView
+from ..view.LineNumbersArea import LineNumbersArea
+from ..view.AtfEditArea import AtfEditArea
+from ..view.SyntaxHighlighter import SyntaxHighlighter
 
 
 class AtfAreaController(object):
@@ -26,12 +30,24 @@ class AtfAreaController(object):
     Creates the ATF area (edit/object) view and handles its actions.
     '''
     def __init__(self, mainControler):
+        # Create text edition area
+        self.edit_area = AtfEditArea(self)
+        # Create text panel to display the line numbers
+        self.line_numbers_area = LineNumbersArea()
         # Create view with a reference to its controller to handle events
         self.view = AtfAreaView(self)
         # Will also need delegating to parent presenter
         self.controller = mainControler
         # Get a reference to the view's undo_manager
         self.undo_manager = self.view.undo_manager
+        # Initialise validation errors
+        self.validation_errors = {}
+        # Needed by syntax highlighter
+        self.edit_area_styledoc = self.edit_area.getStyledDocument()
+        self.line_numbers_styledoc = self.line_numbers_area.getStyledDocument()
+        # Syntax highlighting
+        self.syntax_highlighter = SyntaxHighlighter(self)
+#       self.syntax_highlighter = SyntaxHighlighter(self.line_numbers_styledoc)
 
     def setAtfAreaText(self, text):
         '''
@@ -42,15 +58,15 @@ class AtfAreaController(object):
         INSERT event, so we have to manually group those two together.
         '''
         with self.view.edit_listener.force_compound():
-            self.view.editArea.setText(text)
-            self.view.syntax_highlight()
+            self.view.edit_area.setText(text)
+            self.syntax_highlight()
         self.update_line_numbers()
 
     def getAtfAreaText(self):
         '''
         Short hand for getting Nammu's text area's content.
         '''
-        return self.view.editArea.getText()
+        return self.view.edit_area.getText()
 
     def clearAtfArea(self):
         '''
@@ -58,29 +74,22 @@ class AtfAreaController(object):
         pile, repaint the line numbers and remove the styling from previous
         validation highlight.
         '''
-        self.view.editArea.setText("")
+        self.view.edit_area.setText("")
         # When opening a new file we should discard the previous edits
         self.view.undo_manager.discardAllEdits()
         # Reload line numbers text panel
-        self.view.repaint_line_numbers(0)
+        self.repaint_line_numbers(0)
         # Clear tooltips
         self.clearToolTips()
         # Clear validation errors
-        self.clear_validation_errors()
-
-    def clear_validation_errors(self):
-        '''
-        Clear validation errors not to inherit wrong styling from previously
-        open files.
-        '''
-        self.view.validation_errors = {}
+        self.validation_errors = {}
 
     def clearToolTips(self):
         '''
         We don't want tooltips from previous validations appearing after the
         file has been re-validated or another file has been opened.
         '''
-        self.view.editArea.setToolTipText(None)
+        self.view.edit_area.setToolTipText(None)
 
     def update_line_numbers(self):
         '''
@@ -89,7 +98,13 @@ class AtfAreaController(object):
         # Get how many lines are in the file
         n_lines = self.getAtfAreaText().count('\n')
         # Reload line numbers text panel
-        self.view.repaint_line_numbers(n_lines)
+        self.repaint_line_numbers(n_lines)
+
+    def set_validation_errors(self, validation_errors):
+        '''
+        Short hand for refreshing validation errors in ATF area.
+        '''
+        self.validation_errors = validation_errors
 
     def undo(self):
         '''
@@ -124,4 +139,19 @@ class AtfAreaController(object):
         Calls to copy, paste and cut methods are just passed to text area.
         '''
         if name in ('copy', 'paste', 'cut'):
-            return getattr(self.view.editArea, name)
+            return getattr(self.view.edit_area, name)
+
+    def repaint_line_numbers(self, n_lines):
+        '''
+        Draws line numbers in corresponding panel.
+        '''
+        # Create line numbers
+        numbers = ""
+        for line in range(n_lines + 1):
+            numbers += str(line + 1) + ": \n"
+
+        # Print in line numbers' area
+        self.line_numbers_area.setText(numbers)
+
+    def syntax_highlight(self):
+        self.syntax_highlighter.syntax_highlight()
