+++ python/nammu/utils/NammuConsoleHandler.py
@@ -31,7 +31,7 @@ class NammuConsoleHandler(StreamHandler):
         Needs a reference to nammu to be able to get ahold of Nammu's console.
         """
         super(logging.StreamHandler, self).__init__()
-        # Fixes an innocuous but ugly error message when existing Nammu.
+        # Fixes an innocuous but ugly error message when exiting Nammu.
         self.stream = None
         self.nammu_console = nammu_console
